functions_python/requirements.txt
numpy>=1.23.0
opencv-python>=4.8.0
scipy>=1.10.0
functions-framework>=3.4.0
Pillow>=10.0.0
boto3>=1.28.0
pandas>=2.0.0
tqdm>=4.65.0
kaggle>=1.5.16
matplotlib>=3.7.0
scikit-learn>=1.3.0
flask>=2.3.0
werkzeug>=2.3.0